wechat-bot/requirements.txt
requests
pyyaml
openai
playwright
beautifulsoup4
notion-client
schedule
loguru
